Rename download_scm to download_svn and just care about svn
The distributed systems like git or hg work differently which
makes the original idea a bit clumsy (unfortunately
git archive --remote=$REPO_URL $REVISION doesn't work reliable
because it needs to be explicitely activated in the git server
configuration.

On the upside these modern scm systems normaly provide a web
interface where you can download a tarball of a given revision.
So we will wget from those if we need to.

Signed-off-by: Marc Andre Tanner <mat@brain-dump.org>